Split Electron acceptance scenarios
